backend/requirements.txt
# Core Django + API
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.3.1

# HTTP + HTML parsing
requests==2.31.0
beautifulsoup4==4.12.3

# NLP + tokenization
nltk==3.8.1

# Embedding + vector database
sentence-transformers==2.2.2
chromadb==0.4.22

# PyTorch for sentence-transformers
torch==2.2.2

# Fix for Transformers with Keras 3 (backwards-compatible tf-keras)
tf-keras==2.15.0
